backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.31.0
pyjwt==2.8.0
python-multipart==0.0.6
python-binance==1.0.19